Implement API's custom classes
This adds support for Questionmark's Dummy, Fallback Dummy and Old Dummy
GTINs.

To test this I've added a test for every valid GTIN class to assert the
given GTIN parses to the correct class (relevant for typing and
meta-data purposes) and that if we break the check digit, that we still
get the correct class and detect it's no longer valid.